@@ -1,14 +1,14 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.2.51'
+    ext.kotlin_version = '1.2.61'
 
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.3'
+        classpath 'com.android.tools.build:gradle:3.2.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         classpath 'com.github.dcendents:android-maven-gradle-plugin:2.1'
         // NOTE: Do not place your application dependencies here; they belong
@@ -24,16 +24,16 @@ allprojects {
 }
 
 
-subprojects {
-    project.configurations.all {
-        resolutionStrategy.eachDependency { details ->
-            if (details.requested.group == 'com.android.support'
-                    && !details.requested.name.contains('multidex') ) {
-                details.useVersion  "27.1.1"
-            }
-        }
-    }
-}
+//subprojects {
+//    project.configurations.all {
+//        resolutionStrategy.eachDependency { details ->
+//            if (details.requested.group == 'com.android.support'
+//                    && !details.requested.name.contains('multidex') ) {
+//                details.useVersion  "27.1.1"
+//            }
+//        }
+//    }
+//}
 
 
 task clean(type: Delete) {
@@ -41,16 +41,16 @@ task clean(type: Delete) {
 }
 
 ext {
-    compileSdkVersion = 27
+    compileSdkVersion = 28
     minSdkVersion = 17
     targetSdkVersion = 27
 
     retrofit2Version = '2.4.0'
     androidSupportVersion = '27.1.1'
+    androidXVersion = '1.0.0'
 
 
-
-    glideVersion = '4.7.1'
+    glideVersion = '4.8.0'
 
     leakcanaryVersion = '1.5.1'
 
