@@ -1,7 +1,7 @@
 package com.jarvanmo.kal.chooser;
 
-import android.support.annotation.NonNull;
-import android.support.v4.app.DialogFragment;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.DialogFragment;
 
 public interface OnChooseListener {
      void onChoose(@NonNull DialogFragment dialogFragment,@NonNull Object content);
